🔌 (#7) - add controll progeress type component
